src/requirements.txt
polars
streamlit
yfinance
numpy
pdfplumber
plotly.express